Refactor appcontroller files
